requirements.txt
pyserial==3.4.0